+++ backend/src/baserow/contrib/database/api/rows/views.py
@@ -9,6 +9,7 @@ from drf_spectacular.utils import extend_schema
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.request import Request
 from rest_framework.response import Response
+from rest_framework.status import HTTP_204_NO_CONTENT
 from rest_framework.views import APIView
 
 from baserow.api.decorators import (
@@ -40,6 +41,7 @@ from baserow.contrib.database.api.rows.errors import (
     ERROR_ROW_IDS_NOT_UNIQUE,
 )
 from baserow.contrib.database.api.rows.serializers import (
+    GetRowAdjacentSerializer,
     example_pagination_row_serializer_class,
 )
 from baserow.contrib.database.api.tables.errors import ERROR_TABLE_DOES_NOT_EXIST
@@ -267,7 +269,11 @@ class RowsView(APIView):
             ),
             401: get_error_schema(["ERROR_NO_PERMISSION_TO_TABLE"]),
             404: get_error_schema(
-                ["ERROR_TABLE_DOES_NOT_EXIST", "ERROR_FIELD_DOES_NOT_EXIST"]
+                [
+                    "ERROR_TABLE_DOES_NOT_EXIST",
+                    "ERROR_FIELD_DOES_NOT_EXIST",
+                    "ERROR_VIEW_DOES_NOT_EXIST",
+                ]
             ),
         },
     )
@@ -1236,3 +1242,140 @@ class BatchDeleteRowsView(APIView):
         )
 
         return Response(status=204)
+
+
+class RowAdjacentView(APIView):
+    permission_classes = (IsAuthenticated,)
+
+    @extend_schema(
+        parameters=[
+            OpenApiParameter(
+                name="table_id",
+                location=OpenApiParameter.PATH,
+                type=OpenApiTypes.INT,
+                description="Returns the row of the table related to the provided "
+                "value.",
+            ),
+            OpenApiParameter(
+                name="row_id",
+                location=OpenApiParameter.PATH,
+                type=OpenApiTypes.INT,
+                description="Returns the row adjacent the provided value.",
+            ),
+            OpenApiParameter(
+                name="view_id",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.INT,
+                description="Applies the filters and sorts of the provided view.",
+            ),
+            OpenApiParameter(
+                name="user_field_names",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.BOOL,
+                description=(
+                    "A flag query parameter which if provided the returned json "
+                    "will use the user specified field names instead of internal "
+                    "Baserow field names (field_123 etc). "
+                ),
+            ),
+            OpenApiParameter(
+                name="previous",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.BOOL,
+                description=(
+                    "A flag query parameter which if provided returns the"
+                    "previous row to the specified row_id. If it's not set"
+                    "it will return the next row."
+                ),
+            ),
+            OpenApiParameter(
+                name="search",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.STR,
+                description="If provided, the adjacent row will be one that matches"
+                "the search query.",
+            ),
+        ],
+        tags=["Database table rows"],
+        operation_id="get_database_table_row",
+        description=(
+            "Fetches the adjacent row to a given row_id in the table with the "
+            "given table_id. If the previous flag is set it will return the "
+            "previous row, otherwise it will return the next row. You can specify"
+            "a view_id and it will apply the filters and sorts of the provided "
+            "view."
+        ),
+        responses={
+            200: get_example_row_serializer_class(
+                example_type="get", user_field_names=True
+            ),
+            204: None,
+            400: get_error_schema(
+                ["ERROR_USER_NOT_IN_GROUP", "ERROR_REQUEST_BODY_VALIDATION"]
+            ),
+            404: get_error_schema(
+                [
+                    "ERROR_TABLE_DOES_NOT_EXIST",
+                    "ERROR_ROW_DOES_NOT_EXIST",
+                    "ERROR_VIEW_DOES_NOT_EXIST",
+                ]
+            ),
+        },
+    )
+    @map_exceptions(
+        {
+            UserNotInGroup: ERROR_USER_NOT_IN_GROUP,
+            TableDoesNotExist: ERROR_TABLE_DOES_NOT_EXIST,
+            ViewDoesNotExist: ERROR_VIEW_DOES_NOT_EXIST,
+            RowDoesNotExist: ERROR_ROW_DOES_NOT_EXIST,
+        }
+    )
+    @validate_query_parameters(GetRowAdjacentSerializer)
+    def get(
+        self, request: Request, table_id: int, row_id: int, query_params: Dict[str, Any]
+    ) -> Response:
+
+        previous = query_params.get("previous")
+        view_id = query_params.get("view_id")
+        user_field_names = query_params.get("user_field_names")
+        search = query_params.get("search")
+
+        table = TableHandler().get_table(table_id)
+        table.database.group.has_user(request.user, raise_error=True)
+
+        model = table.get_model()
+        queryset = model.objects.all().enhance_by_fields()
+
+        if search is not None:
+            queryset = queryset.search_all_fields(search)
+
+        view = None
+        if view_id:
+            view_handler = ViewHandler()
+            view = view_handler.get_view(view_id)
+
+            if view.table_id != table.id:
+                raise ViewDoesNotExist()
+
+            queryset = view_handler.apply_filters(view, queryset)
+            queryset = view_handler.apply_sorting(view, queryset)
+
+        try:
+            row = queryset.get(pk=row_id)
+        except model.DoesNotExist:
+            raise RowDoesNotExist(row_id)
+
+        adjacent_row = RowHandler().get_adjacent_row(
+            row, queryset, previous=previous, view=view
+        )
+
+        # Don't fail, just let the user know there isn't an adjacent row
+        if adjacent_row is None:
+            return Response(status=HTTP_204_NO_CONTENT)
+
+        serializer_class = get_row_serializer_class(
+            model, RowSerializer, is_response=True, user_field_names=user_field_names
+        )
+        serializer = serializer_class(adjacent_row)
+
+        return Response(serializer.data)