+++ web-frontend/modules/database/components/table/Table.vue
@@ -139,11 +139,18 @@
         :table="table"
         :view="view"
         :fields="fields"
-        :row="row"
         :read-only="readOnly"
         :store-prefix="storePrefix"
         @refresh="refresh"
         @selected-row="$emit('selected-row', $event)"
+        @navigate-previous="
+          (row, activeSearchTerm) =>
+            $emit('navigate-previous', row, activeSearchTerm)
+        "
+        @navigate-next="
+          (row, activeSearchTerm) =>
+            $emit('navigate-next', row, activeSearchTerm)
+        "
       />
       <div v-if="viewLoading" class="loading-overlay"></div>
     </div>
@@ -207,11 +214,6 @@ export default {
     view: {
       required: true,
       validator: (prop) => typeof prop === 'object' || prop === undefined,
-    },
-    row: {
-      type: Object,
-      required: false,
-      default: null,
     },
     tableLoading: {
       type: Boolean,