+++ changelog.md
@@ -66,6 +66,8 @@ For example:
 * Added Multiple Collaborators field type. [#1119](https://gitlab.com/bramw/baserow/-/issues/1119)
 * Added missing success printouts to `count_rows` and `calculate_storage_usage` commands.
 * Add `isort` settings to sort python imports.
+* Add row url parameter to `gallery` and `kanban` view.
+* Add navigation buttons to the `RowEditModal`.
 * Introduced a premium form survey style theme. [#524](https://gitlab.com/bramw/baserow/-/issues/524).
 * Allow creating new rows when selecting a related row [#1064](https://gitlab.com/bramw/baserow/-/issues/1064).
 * Add row url parameter to `gallery` and `kanban` view.